a2a/my-project/pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "NAZRUL Islam", email = "nislam@us.ibm.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "google-a2a",
    "python-dotenv>=1.1.0",
]

[project.scripts]
my-project = "my_project:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
google-a2a = { git = "https://github.com/djsamseng/A2A", subdirectory = "samples/python", branch = "prefixPythonPackage" }
